requirements.txt
nltk
numpy
flask
torch
torchvision
pillow